Fixes setExtension of TipImage.java in org.eclipse.tips.tests #525
In this commit the tests setExtension and setExtension2 are reactivated and changed. For this the method setExtension of TipImage.java is corrected to properly change the extension. Now if setExtension is called, not only the field fExtension is updated but also the field fBase64Image. Mind that for this change the field fBase64Image can't be final. Contributes to #525.