Implement move and collapse operations and change delete for archives
the collapse operation deletes the linked folder by calling the function with the new COLLAPSE flag indicating that the Archive file should not be deleted.

The normal delete mechanism is adjusted so that is only deletes the Archvie file in the local file system when the COLLAPSE flag is not set.

The move implementation collapses the linked folder, then moves the archvie file and then expands the file in the new location.