Move ZipExpander and ZipCollapser to platform projects
This commit migrates the ZipCollapser and ZipExpander classes from package org.eclipse.ui.handlers of org.eclipse.ui.workbench into org.eclipse.core.resources so that the functionality for expanding/collapsing archive is not in ui projects anymore. The handlers are still in ui.workbench and just call ZipExpander/ZipCollapser.